requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
pdfplumber>=0.9.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0
spacy>=3.6.0
nltk>=3.8.1
scikit-learn>=1.3.0
Pillow>=10.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
requests>=2.31.0
streamlit-option-menu>=0.3.6
streamlit-chat>=0.1.1
